refactor: Component 파일 분리
